requirements.txt
pyautogui==0.9.53
mysqlclient==2.1.1
mysql-connector-python==8.0.32
pillow==9.5.0
pytest==7.3.1 